# Promtail: Improve LabelsMap CPU performance