--- a/clients/pkg/promtail/client/batch.go
+++ b/clients/pkg/promtail/client/batch.go
@@ -96,9 +96,8 @@ func labelsMapToString(ls model.LabelSet, without model.LabelName) string {
 		}
 
 		b.WriteString(string(l))
-		b.WriteString(`="`)
+		b.WriteString(`=`)
 		b.WriteString(strconv.Quote(string(ls[l])))
-		b.WriteString(`"`)
 	}
 	b.WriteByte('}')