--- a/clients/pkg/promtail/client/batch_test.go
+++ b/clients/pkg/promtail/client/batch_test.go
@@ -172,7 +172,7 @@ var result string
 func BenchmarkLabelsMapToString(b *testing.B) {
 	labelSet := make(model.LabelSet)
 	labelSet["label"] = "value"
-	labelSet["label1"] = "value"
+	labelSet["label1"] = "value2"
 	labelSet["label2"] = "value3"
 	labelSet["__tenant_id__"] = "another_value"