**What this PR does / why we need it**:

The labelsMapToString function takes a relative high amount of CPU in a simple promtail pipeline. By replacing the fmt.Sprintf with a string. Builder we achieve a 50+% improvement in the function.  In my local promtail installation I get about a 10% global cpu reduction. 

```
name                 old time/op    new time/op    delta
LabelsMapToString-8    1.10µs ± 3%    0.30µs ± 4%  -72.40%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
LabelsMapToString-8      344B ± 0%      184B ± 0%  -46.51%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
LabelsMapToString-8      14.0 ± 0%       6.0 ± 0%  -57.14%  (p=0.008 n=5+5)
```

**Special notes for your reviewer**:
The code does get more complex, and this might be a trade-off. Considering promtail is running on many machines (ie. kubernetes nodes) I think the gain is worth it.

Two alternatives:
- Only replace the `lstrs = append(lstrs, fmt.Sprintf("%s=%q", l, v))` call, as it takes most of the time
- Get rid of the explicit `b.Grow` calls and the tracking of size. This only adds a slight improvement

**Checklist**
- [x] Reviewed the `CONTRIBUTING.md` guide
- [ ] Documentation added
- [x] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`